feat: apply changed event rule design
